Take advantage of broadcasting/implicit expansion in chebtech.clenshaw().
This has two benefits:

 - There is a measurable speedup when evaluating a high-degree
   array-valued chebtech with many columns at many points.

 - It is no longer necessary to maintain duplicate versions of the code
   for the scalar and vector cases.

To expand on the second:  with implicit expansion, the *only* difference
between the scalar and vector codes is the indexing of the coefficient
array as c(k) vs. c(k,:).  I am unable to measure a speed difference
between these for the case when c has only one column.

This breaks compatibility with MATLAB versions older than R2016b.